Accounts/pyproject.toml
[tool.poetry]
name = "accounts"
version = "0.1.0"
description = "Main User System"
authors = ["nobinkhan <nazrulkhan.dk@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
django-countries = "^7.5"
djangorestframework-simplejwt = "^5.2.2"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.0.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
